requirements.txt
conda==4.3.16
matplotlib==3.4.3
pandas==1.4.3
numba==0.53.0
numpy==1.21.4
scipy==1.6.3
Pillow==9.2.0
plotly==5.10.0
pycaret==2.2.3
seaborn==0.11.2
shap==0.41.0
streamlit==1.11.1
tune-sklearn==0.4.1
scikit-plot==0.3.7
xgboost==1.6.1
urllib3==1.24.3
requests==2.28.1
flask==1.1.1
jinja2==2.11.0
